requirements.txt
Django>=4.2.7
Pillow>=10.0.0
django-modeltranslation>=0.18.11
django-ckeditor>=6.7.0
django-admin-thumbnails>=0.2.8
django-meta>=2.4.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
